+++ apps/client/public/locales/ru/translation.json
@@ -64,6 +64,7 @@
     "enter": "войти"
   },
   "room": {
+    "title": "Комната",
     "exit-button": "Покинуть комнату"
   }
 }
